Adicionados Get e Put

+++ src/main/java/br/com/etec/cinehas/repository/SalaRepository.java
@@ -3,7 +3,7 @@ package br.com.etec.cinehas.repository;
 import br.com.etec.cinehas.entity.Sala;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SalaRepository extends JpaRepository<Sala, Long>
+public interface SalaRepository extends JpaRepository<Sala, Integer>
 {
 
 }
